% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_utils.R
\name{.check_data_quality}
\alias{.check_data_quality}
\title{Data Quality Check}
\usage{
.check_data_quality(data, target, factors)
}
\arguments{
\item{data}{Data frame}

\item{target}{Target variable name}

\item{factors}{Factor variable names}
}
\value{
Data quality check report list
}
\description{
Checks if data is suitable for modeling, returns check report
}
\keyword{internal}
