% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_utils.R
\docType{data}
\name{.MODEL_CONFIG}
\alias{.MODEL_CONFIG}
\title{Global Configuration Constants}
\format{
An object of class \code{list} of length 2.
}
\usage{
.MODEL_CONFIG
}
\description{
Thresholds and configuration for predictive modeling
}
\keyword{internal}
