% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_utils.R
\name{.progress_message}
\alias{.progress_message}
\title{User-friendly Progress Message}
\usage{
.progress_message(stage, details = NULL)
}
\arguments{
\item{stage}{Stage name}

\item{details}{Details information}
}
\description{
Outputs formatted progress information
}
\keyword{internal}
