% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictiveAnalyzer.R
\name{PredictiveAnalyzer}
\alias{PredictiveAnalyzer}
\title{PredictiveAnalyzer: Quality Prediction Modeling Analysis Executor}
\description{
Core executor for prediction modeling tasks.
Coordinates the complete workflow of model training, diagnostic analysis,
and explainability analysis. Adopts "Business Group / Expert" dual mode:
\itemize{
\item Business Group: Intelligent defaults, one-click results
\item Expert: Full control, supports expert customization
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_trainer}}{ModelTrainer instance}

\item{\code{diagnostic_analyzer}}{DiagnosticAnalyzer instance}

\item{\code{explainability_engine}}{ExplainabilityEngine instance}

\item{\code{results}}{Standardized results container}

\item{\code{params}}{Parameter list}

\item{\code{model_trainer}}{Model trainer}

\item{\code{diagnostic_analyzer}}{Diagnostic analyzer}

\item{\code{explainability_engine}}{Explainability engine}

\item{\code{results}}{Standardized results container}

\item{\code{params}}{Parameter list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictiveAnalyzer-new}{\code{PredictiveAnalyzer$new()}}
\item \href{#method-PredictiveAnalyzer-reset}{\code{PredictiveAnalyzer$reset()}}
\item \href{#method-PredictiveAnalyzer-set_statistic}{\code{PredictiveAnalyzer$set_statistic()}}
\item \href{#method-PredictiveAnalyzer-set_diagnostic}{\code{PredictiveAnalyzer$set_diagnostic()}}
\item \href{#method-PredictiveAnalyzer-set_datatable}{\code{PredictiveAnalyzer$set_datatable()}}
\item \href{#method-PredictiveAnalyzer-set_raw_output}{\code{PredictiveAnalyzer$set_raw_output()}}
\item \href{#method-PredictiveAnalyzer-get_results}{\code{PredictiveAnalyzer$get_results()}}
\item \href{#method-PredictiveAnalyzer-run}{\code{PredictiveAnalyzer$run()}}
\item \href{#method-PredictiveAnalyzer-predict_new}{\code{PredictiveAnalyzer$predict_new()}}
\item \href{#method-PredictiveAnalyzer-clone}{\code{PredictiveAnalyzer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictiveAnalyzer-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictiveAnalyzer-new}{}}}
\subsection{Method \code{new()}}{
Initialize analyzer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictiveAnalyzer$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictiveAnalyzer-reset"></a>}}
\if{latex}{\out{\hypertarget{method-PredictiveAnalyzer-reset}{}}}
\subsection{Method \code{reset()}}{
Reset results container
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictiveAnalyzer$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictiveAnalyzer-set_statistic"></a>}}
\if{latex}{\out{\hypertarget{method-PredictiveAnalyzer-set_statistic}{}}}
\subsection{Method \code{set_statistic()}}{
Store single statistic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictiveAnalyzer$set_statistic(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Statistic name}

\item{\code{value}}{Statistic value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictiveAnalyzer-set_diagnostic"></a>}}
\if{latex}{\out{\hypertarget{method-PredictiveAnalyzer-set_diagnostic}{}}}
\subsection{Method \code{set_diagnostic()}}{
Store single diagnostic entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictiveAnalyzer$set_diagnostic(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Diagnostic name}

\item{\code{value}}{Diagnostic value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictiveAnalyzer-set_datatable"></a>}}
\if{latex}{\out{\hypertarget{method-PredictiveAnalyzer-set_datatable}{}}}
\subsection{Method \code{set_datatable()}}{
Store single data table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictiveAnalyzer$set_datatable(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Table name}

\item{\code{value}}{Data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictiveAnalyzer-set_raw_output"></a>}}
\if{latex}{\out{\hypertarget{method-PredictiveAnalyzer-set_raw_output}{}}}
\subsection{Method \code{set_raw_output()}}{
Store raw output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictiveAnalyzer$set_raw_output(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Raw object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictiveAnalyzer-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-PredictiveAnalyzer-get_results}{}}}
\subsection{Method \code{get_results()}}{
Get standardized results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictiveAnalyzer$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictiveAnalyzer-run"></a>}}
\if{latex}{\out{\hypertarget{method-PredictiveAnalyzer-run}{}}}
\subsection{Method \code{run()}}{
Execute prediction modeling analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictiveAnalyzer$run(data, plan)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Training data}

\item{\code{plan}}{PredictivePlan object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self reference
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictiveAnalyzer-predict_new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictiveAnalyzer-predict_new}{}}}
\subsection{Method \code{predict_new()}}{
Predict new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictiveAnalyzer$predict_new(new_data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{New data}

\item{\code{...}}{Other parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictiveAnalyzer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictiveAnalyzer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictiveAnalyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
